requirements.txt
certifi
discord.py~=1.7.0
Pillow~=8.2.0
aiohttp-basicauth~=0.2.0